First working state having split Match and Explain behaviour out of BoundConcludableResolver into subclasses.
