+++ reasoner/resolution/resolver/BoundConcludableResolver.java
@@ -34,6 +34,7 @@ import com.vaticle.typedb.core.reasoner.resolution.framework.AnswerCache.Subsump
 import com.vaticle.typedb.core.reasoner.resolution.framework.ReiterationQuery;
 import com.vaticle.typedb.core.reasoner.resolution.framework.Request;
 import com.vaticle.typedb.core.reasoner.resolution.framework.RequestState;
+import com.vaticle.typedb.core.reasoner.resolution.framework.RequestState.CachingRequestState;
 import com.vaticle.typedb.core.reasoner.resolution.framework.Resolver;
 import com.vaticle.typedb.core.reasoner.resolution.framework.Response;
 import com.vaticle.typedb.core.traversal.TraversalEngine;
@@ -53,17 +54,16 @@ import java.util.Set;
 import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
 import static com.vaticle.typedb.core.common.iterator.Iterators.iterate;
 
-public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>  {
+public abstract class BoundConcludableResolver extends Resolver<BoundConcludableResolver>  {
 
     private static final Logger LOG = LoggerFactory.getLogger(BoundConcludableResolver.class);
-    private AnswerCache<?, ConceptMap> cache;
-    private final Concludable concludable;
-    private final Set<Identifier.Variable.Retrievable> unboundVars;
-    private final ConceptMap bounds;
     private final Map<Driver<ConclusionResolver>, Rule> resolverRules;
     private final LinkedHashMap<Driver<ConclusionResolver>, Set<Unifier>> applicableRules;
-    private final Map<Request, RequestState.CachingRequestState<?, ConceptMap>> requestStates;
-    private RequestState exploringRequestState;
+    private final Map<Request, CachingRequestState<?, ConceptMap>> requestStates;
+    protected final Concludable concludable;
+    protected final ConceptMap bounds;
+    protected AnswerCache<?, ConceptMap> cache;
+    protected RequestState exploringRequestState;
 
     public BoundConcludableResolver(Driver<BoundConcludableResolver> driver, Concludable concludable, ConceptMap bounds,
                                     Map<Driver<ConclusionResolver>, Rule> resolverRules,
@@ -72,11 +72,10 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
                                     ConceptManager conceptMgr, boolean resolutionTracing) {
         super(driver, initName(concludable, bounds), registry, traversalEngine, conceptMgr, resolutionTracing);
         this.concludable = concludable;
-        this.unboundVars = unboundVars(concludable.pattern());
         this.bounds = bounds;
         this.resolverRules = resolverRules;
         this.applicableRules = applicableRules;
-        this.cache = new ConceptMapCache(new HashMap<>(), bounds, () -> traversalIterator(concludable.pattern(), bounds));
+        this.cache = createCache(concludable.pattern(), bounds);
         this.requestStates = new HashMap<>();
         this.exploringRequestState = null;
     }
@@ -85,6 +84,8 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
         return BoundConcludableResolver.class.getSimpleName() + "(pattern: " + concludable.pattern() + " bounds: " +
                 bounds.toString() + ")";
     }
+
+    abstract AnswerCache<?, ConceptMap> createCache(Conjunction conjunction, ConceptMap bounds);
 
     @Override
     public void receiveRequest(Request fromUpstream, int iteration) {
@@ -126,17 +127,23 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
     }
 
     private void receiveDirectRequest(Request fromUpstream, int iteration) {
-//        sendAnswerOrFail(fromUpstream, iteration, requestStates.computeIfAbsent(
-//                fromUpstream, request -> new BoundRetrievableResolver.BoundRequestState(request, cache, iteration)));
-        sendAnswerOrSearchRulesOrFail(fromUpstream, iteration, requestStates.computeIfAbsent(
-                fromUpstream, request -> createRequestState(fromUpstream, iteration)));
+        assert fromUpstream.partialAnswer().isConcludable();
+        CachingRequestState<?, ConceptMap> requestState = requestStates.computeIfAbsent(
+                fromUpstream, request -> createRequestState(fromUpstream, iteration));
+        if (exploringRequestState == null) {
+            assert requestState.isExploration();
+            exploringRequestState = requestState;
+        }
+        sendAnswerOrSearchRulesOrFail(fromUpstream, iteration, requestState);
     }
+
+    abstract CachingRequestState<?, ConceptMap> createRequestState(Request fromUpstream, int iteration);
 
     @Override
     protected void receiveAnswer(Response.Answer fromDownstream, int iteration) {
         if (isTerminated()) return;
         Request fromUpstream = fromUpstream(fromDownstream.sourceRequest());
-        RequestState.CachingRequestState<?, ConceptMap> requestState = this.requestStates.get(fromUpstream);
+        CachingRequestState<?, ConceptMap> requestState = this.requestStates.get(fromUpstream);
         assert requestState.isExploration();
         requestState.asExploration().newAnswer(fromDownstream.answer());
         sendAnswerOrSearchRulesOrFail(fromUpstream, iteration, requestState);
@@ -150,7 +157,7 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
         Request toDownstream = fromDownstream.sourceRequest();
         Request fromUpstream = fromUpstream(toDownstream);
 
-        RequestState.CachingRequestState<?, ConceptMap> requestState = this.requestStates.get(fromUpstream);
+        CachingRequestState<?, ConceptMap> requestState = this.requestStates.get(fromUpstream);
         assert iteration == requestState.iteration();
         if (requestState.isExploration()) {
             requestState.asExploration().downstreamManager().removeDownstream(fromDownstream.sourceRequest());
@@ -158,7 +165,7 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
         sendAnswerOrSearchRulesOrFail(fromUpstream, iteration, requestState);
     }
 
-    private void sendAnswerOrSearchRulesOrFail(Request fromUpstream, int iteration, RequestState.CachingRequestState<?, ConceptMap> requestState) {
+    private void sendAnswerOrSearchRulesOrFail(Request fromUpstream, int iteration, CachingRequestState<?, ConceptMap> requestState) {
         Optional<AnswerState.Partial.Compound<?, ?>> upstreamAnswer = requestState.nextAnswer().map(AnswerState.Partial::asCompound);
         if (upstreamAnswer.isPresent()) {
             /*
@@ -189,7 +196,6 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
         }
     }
 
-
     private void requestFromSubsumer(Request.ToSubsumed fromUpstream, int iteration) {
         Request toSubsumer = Request.ToSubsumer.create(driver(), fromUpstream.subsumer(),
                                                        fromUpstream, fromUpstream.partialAnswer());
@@ -215,75 +221,7 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
         requestStates.clear();
     }
 
-    private static Set<Identifier.Variable.Retrievable> unboundVars(Conjunction conjunction) {
-        Set<Identifier.Variable.Retrievable> missingBounds = new HashSet<>();
-        iterate(conjunction.variables()).filter(var -> var.id().isRetrievable()).forEachRemaining(var -> {
-            if (var.isType() && !var.asType().label().isPresent()) missingBounds.add(var.asType().id().asRetrievable());
-            else if (var.isThing() && !var.asThing().iid().isPresent())
-                missingBounds.add(var.asThing().id().asRetrievable());
-        });
-        return missingBounds;
-    }
-
-    protected RequestState.CachingRequestState<?, ConceptMap> createRequestState(Request fromUpstream, int iteration) {
-        LOG.debug("{}: Creating new Responses for iteration{}, request: {}", name(), iteration, fromUpstream);
-        assert fromUpstream.partialAnswer().isConcludable();
-        RequestState.CachingRequestState<?, ConceptMap> requestState;
-        if (fromUpstream.partialAnswer().asConcludable().isExplain()) {
-            requestState = createExplainRequestState(fromUpstream, iteration);
-        } else if (fromUpstream.partialAnswer().asConcludable().isMatch()) {
-            requestState = createMatchRequestState(fromUpstream, iteration);
-        } else {
-            throw TypeDBException.of(ILLEGAL_STATE);
-        }
-        if (exploringRequestState == null) {
-            assert requestState.isExploration();
-            exploringRequestState = requestState;
-        }
-        return requestState;
-    }
-
-    protected RequestState.CachingRequestState<?, ConceptMap> createExplainRequestState(Request fromUpstream, int iteration) {
-        RequestState.CachingRequestState<?, ConceptMap> requestState;
-        if (exploringRequestState != null) {
-            if (cache.isConceptMapCache()) {
-                // We have a cache already which we must evict to use a cache suitable for explaining
-                cache = new AnswerCache.ConcludableAnswerCache(new HashMap<>(), bounds);
-                requestState = new Explain(fromUpstream, cache.asConcludableAnswerCache(), iteration);
-                requestState.asExploration().downstreamManager().addDownstreams(ruleDownstreams(fromUpstream));
-            } else if (cache.isConcludableAnswerCache()) {
-                requestState = new FollowingExplain(fromUpstream, cache.asConcludableAnswerCache(), iteration);
-            } else {
-                throw TypeDBException.of(ILLEGAL_STATE);
-            }
-        } else {
-            cache = new AnswerCache.ConcludableAnswerCache(new HashMap<>(), bounds);
-            requestState = new Explain(fromUpstream, cache.asConcludableAnswerCache(), iteration);
-            requestState.asExploration().downstreamManager().addDownstreams(ruleDownstreams(fromUpstream));
-        }
-        return requestState;
-    }
-
-    protected RequestState.CachingRequestState<?, ConceptMap> createMatchRequestState(Request fromUpstream, int iteration) {
-        RequestState.CachingRequestState<?, ConceptMap> requestState;
-        if (exploringRequestState != null) {
-            if (cache.isConceptMapCache()) {
-                requestState = new FollowingMatch(fromUpstream, cache.asConceptMapCache(), iteration);
-            } else if (cache.isConcludableAnswerCache()) {
-                requestState = new FollowingExplain(fromUpstream, cache.asConcludableAnswerCache(), iteration);
-            } else {
-                throw TypeDBException.of(ILLEGAL_STATE);
-            }
-        } else {
-            cache = new ConceptMapCache(new HashMap<>(), bounds, () -> traversalIterator(concludable.pattern(), bounds));
-            boolean singleAnswerRequired = bounds.concepts().keySet().containsAll(unboundVars);
-            requestState = new Match(fromUpstream, cache.asConceptMapCache(), iteration, singleAnswerRequired);
-            requestState.asExploration().downstreamManager().addDownstreams(ruleDownstreams(fromUpstream));
-        }
-        return requestState;
-    }
-
-    private List<Request> ruleDownstreams(Request fromUpstream) {
+    protected List<Request> ruleDownstreams(Request fromUpstream) {
         // loop termination: when receiving a new request, we check if we have seen it before from this root query
         // if we have, we do not allow rules to be registered as possible downstreams
         List<Request> downstreams = new ArrayList<>();
@@ -300,116 +238,4 @@ public class BoundConcludableResolver extends Resolver<BoundConcludableResolver>
         return downstreams;
     }
 
-    private class FollowingMatch extends RequestState.CachingRequestState<ConceptMap, ConceptMap> {
-
-        public FollowingMatch(Request fromUpstream, AnswerCache<ConceptMap, ConceptMap> answerCache, int iteration) {
-            super(fromUpstream, answerCache, iteration, true, true);
-        }
-
-        @Override
-        protected FunctionalIterator<? extends AnswerState.Partial<?>> toUpstream(ConceptMap conceptMap) {
-            return Iterators.single(fromUpstream.partialAnswer().asConcludable().asMatch().toUpstreamLookup(
-                    conceptMap, concludable.isInferredAnswer(conceptMap)));
-        }
-
-    }
-
-    private class Match extends RequestState.CachingRequestState<ConceptMap, ConceptMap> implements RequestState.Exploration {
-
-        private final DownstreamManager downstreamManager;
-        private final boolean singleAnswerRequired;
-
-        public Match(Request fromUpstream, AnswerCache<ConceptMap, ConceptMap> answerCache,
-                     int iteration, boolean singleAnswerRequired) {
-            super(fromUpstream, answerCache, iteration, true, false);
-            this.downstreamManager = new DownstreamManager();
-            this.singleAnswerRequired = singleAnswerRequired;
-        }
-
-        @Override
-        public boolean isExploration() {
-            return true;
-        }
-
-        @Override
-        public Exploration asExploration() {
-            return this;
-        }
-
-        @Override
-        public DownstreamManager downstreamManager() {
-            return downstreamManager;
-        }
-
-        @Override
-        public void newAnswer(AnswerState.Partial<?> partial) {
-            if (!answerCache.isComplete()) answerCache.add(partial.conceptMap());
-        }
-
-        @Override
-        public boolean singleAnswerRequired() {
-            return singleAnswerRequired;
-        }
-
-        @Override
-        protected FunctionalIterator<? extends AnswerState.Partial<?>> toUpstream(ConceptMap conceptMap) {
-            return Iterators.single(fromUpstream.partialAnswer().asConcludable().asMatch().toUpstreamLookup(
-                    conceptMap, concludable.isInferredAnswer(conceptMap)));
-        }
-    }
-
-    private static class FollowingExplain extends RequestState.CachingRequestState<AnswerState.Partial.Concludable<?>, ConceptMap> {
-
-        public FollowingExplain(Request fromUpstream, AnswerCache<AnswerState.Partial.Concludable<?>, ConceptMap> answerCache,
-                                int iteration) {
-            super(fromUpstream, answerCache, iteration, true, true);
-        }
-
-        @Override
-        protected FunctionalIterator<? extends AnswerState.Partial<?>> toUpstream(AnswerState.Partial.Concludable<?> partial) {
-            return Iterators.single(partial.asExplain().toUpstreamInferred());
-        }
-    }
-
-    private static class Explain extends RequestState.CachingRequestState<AnswerState.Partial.Concludable<?>, ConceptMap> implements RequestState.Exploration {
-
-        private final DownstreamManager downstreamManager;
-
-        public Explain(Request fromUpstream, AnswerCache<AnswerState.Partial.Concludable<?>, ConceptMap> answerCache,
-                       int iteration) {
-            super(fromUpstream, answerCache, iteration, false, false);
-            this.downstreamManager = new DownstreamManager();
-        }
-
-        @Override
-        public boolean isExploration() {
-            return true;
-        }
-
-        @Override
-        public Exploration asExploration() {
-            return this;
-        }
-
-        @Override
-        public DownstreamManager downstreamManager() {
-            return downstreamManager;
-        }
-
-        @Override
-        public void newAnswer(AnswerState.Partial<?> partial) {
-            if (!answerCache.isComplete()) answerCache.add(partial.asConcludable());
-        }
-
-        @Override
-        public boolean singleAnswerRequired() {
-            return false;
-        }
-
-        @Override
-        protected FunctionalIterator<? extends AnswerState.Partial<?>> toUpstream(AnswerState.Partial.Concludable<?> partial) {
-            return Iterators.single(partial.asExplain().toUpstreamInferred());
-        }
-
-    }
 }
